Separated ledger derivatio path modal
